Struct dimensioned::dimensioned::Dim

pub struct Dim<D, V>(pub V, pub PhantomData<D>);

This is the primary struct that users of this library will interact with.

Methods

impl<D, V> Dim<D, V>

const fn new(v: V) -> Dim<D, V>

Construct a new Dim object.

It is recommened to use this only where necessary, and to generally use the constants that ship with unit systems to create Dim objects.

Example

use dimensioned::Dim;
use dimensioned::si::{m, Meter};

let x: Dim<Meter, f64> = Dim::new(3.0);
let y = 3.0*m;
assert_eq!(x, y);

fn map<O, F: FnOnce(V) -> O>(self, f: F) -> Dim<D, O>

Map a Dim<D, V> to Dim<D, O> by applying function f to the contained value

Example

use num::traits::Float;
use dimensioned::si::m;

let x = 3.5*m;
assert_eq!(3.0*m, x.map(Float::trunc) );