Struct dimensioned::dimensioned::PhantomDataStable

pub struct PhantomData<T> where T: ?Sized;

PhantomData<T> allows you to describe that a type acts as if it stores a value of type T, even though it does not. This allows you to inform the compiler about certain safety properties of your code.

Though they both have scary names, PhantomData<T> and "phantom types" are unrelated. 👻👻👻

Examples

When handling external resources over a foreign function interface, PhantomData<T> can prevent mismatches by enforcing types in the method implementations:

use std::marker::PhantomData;
use std::mem;

struct ExternalResource<R> {
   resource_handle: *mut (),
   resource_type: PhantomData<R>,
}

impl<R: ResType> ExternalResource<R> {
    fn new() -> ExternalResource<R> {
        let size_of_res = mem::size_of::<R>();
        ExternalResource {
            resource_handle: foreign_lib::new(size_of_res),
            resource_type: PhantomData,
        }
    }

    fn do_stuff(&self, param: ParamType) {
        let foreign_params = convert_params(param);
        foreign_lib::do_stuff(self.resource_handle, foreign_params);
    }
}

Another example: embedding a PhantomData<T> will inform the compiler that one or more instances of the type T could be dropped when instances of the type itself is dropped, though that may not be apparent from the other structure of the type itself. This is commonly necessary if the structure is using an unsafe pointer like *mut T whose referent may be dropped when the type is dropped, as a *mut T is otherwise not treated as owned.