Trait dimensioned::dimensioned::DimToString

pub trait DimToString: Dimension {
    fn to_string() -> String;
}

This trait allows human-friendly printing of dimensioned objects. It is used to implement the traits in std::fmt.

All types created for a unit system will implement this trait. No other types should implement it.

Required Methods

fn to_string() -> String

Gives a human friendly String representation of a Dimension type.