Trait dimensioned::dimensioned::Dimensionless

pub trait Dimensionless: Dimension { }

The only types that implement this trait are the Unitless types that exist in each unit system. It allows more flexibility when handling specifically objects without dimension.

No other types should implement it.