Trait dimensioned::dimensioned::NotDim

pub trait NotDim { }

This traits is implemented by default for everything that is not Dim. It allows a greater level of generic operator overloading than would be possible otherwise.