Trait dimensioned::peano::DivPeano

pub trait DivPeano<RHS>: Peano {
    type Output;
}

Note that, while we define division, we are operating in a ring, so an error will be thrown unless the numerator is divisible by the denominator

Associated Types

type Output