Module dimensioned::si

The si module provides a unit system for use with SI units.

It will also define derived units, although this is not implemented yet.

Structs

SI

Constants

amp

candela

kelvin

kg

m

mole

one

s

Type Definitions

Ampere

Candela

Kelvin

Kilogram

Meter

Mole

Second

Unitless