Module dimensioned::unit_systems

Modules

cgs
The cgs module provides a unit system for use with Gaussian CGS units. It was generated using the make_units! macro. See its documentation for more information.

mks
The mks module provides a unit system for use with Gaussian MKS units. It was generated using the make_units! macro. See its documentation for more information.

si
The si module provides a unit system for use with SI units. It was generated using the make_units! macro. See its documentation for more information.