Module dimensioned::unit_systems::fps

The foot, pound, second system, using the mass pound as a base unit.

Note: this system is incomplete. More derived units and constants are coming.

Following, we list all of the base units, derived units, and constants that are defined in this unit system.

Base Units

Constant 	Unit 	Print Token 	Dimensionn

SQRTFT 	SqrtFoot 	sqrtft

SQRTLB 	SqrtPound 	sqrtlb

S 	Second 	s 	Time

Derived Units

Constant 	Unit 	Unit Definition 	Dimension

FT 	Foot 	SqrtFoot * SqrtFoot 	Length

LB 	Pound 	SqrtPound * SqrtPound 	Mass

Constants

Name 	Constant 	Value 	Unit 	Dimension

Reexports

pub use self::f64consts::*;

Modules

f32consts
Constants defined for this system

f64consts
Constants defined for this system

Structs

FPS
The struct for this unit system

Type Definitions

Foot

Pound

Second

SqrtFoot

SqrtPound

Unitless