Struct dimensioned::unit_systems::fps::FPS

pub struct FPS<V, U> {
    pub value_unsafe: V,
    pub _marker: PhantomData<U>,
}

The struct for this unit system

Fields
value_unsafe: V

This is the value of whatever type we're giving units. Using it directly bypasses all of the dimensional analysis that having a unit system provides, and should be avoided whenever possible.

If using this member is necessary, it is strongly encouraged to wrap the calculation in a dimensionally-safe interface.

_marker: PhantomData<U>

This member is only temporarily public and so its use is considered unstable. Right now, the only way to create a const with units is with this pattern:

extern crate dimensioned as dim;
use dim::si;

const x: si::Meter<f64> = si::Meter { value_unsafe: 3.4, _marker: std::marker::PhantomData };

Once const_fns is stabilized, that will be able to be replaced with a call to Meter::new and _marker will be made private.

Methods

impl<V, U> FPS<V, U>

fn new(v: V) -> Self

Create a new quantity in the $System unit system