Module dimensioned::unit_systems::mks

The Gaussian MKS unit system

Note: this system is incomplete. More derived units and constants are coming.

Following, we list all of the base units, derived units, and constants that are defined in this unit system.

Base Units

Constant 	Unit 	Print Token 	Dimensionn

SQRTM 	SqrtMeter 	sqrtm

SQRTKG 	SqrtKilogram 	sqrtkg

S 	Second 	s 	Time

Derived Units

Constant 	Unit 	Unit Definition 	Dimension

M 	Meter 	SqrtMeter * SqrtMeter 	Length

KG 	Kilogram 	SqrtKilogram * SqrtKilogram 	Mass

MPS 	MeterPerSecond 	Meter / Second 	Velocity

Constants

Name 	Constant 	Value 	Unit 	Dimension

Reexports

pub use self::f64consts::*;

Modules

f32consts
Constants defined for this system

f64consts
Constants defined for this system

Structs

MKS
The struct for this unit system

Type Definitions

Kilogram

Meter

MeterPerSecond

Second

SqrtKilogram

SqrtMeter

Unitless