Module dimensioned::unit_systems::si

The International System of Units (SI)

Where experimental values are used for constants, the values are obtained from the 2014 values from The NIST Reference on Constants, Unit, and Uncertainty.

Following, we list all of the base units, derived units, and constants that are defined in this unit system.

Base Units

Constant 	Unit 	Print Token 	Dimensionn

M 	Meter 	m 	Length

KG 	Kilogram 	kg 	Mass

S 	Second 	s 	Time

A 	Ampere 	A 	Current

K 	Kelvin 	K 	Temperature

CD 	Candela 	cd 	LuminousIntensity

MOL 	Mole 	mol 	AmountOfSubstance

Derived Units

Constant 	Unit 	Unit Definition 	Dimension

HZ 	Hertz 	Unitless / Second 	Frequency

N 	Newton 	Kilogram * Meter / Second2 	Force

PA 	Pascal 	Newton / Meter2 	Pressure

J 	Joule 	Newton * Meter 	Energy

W 	Watt 	Joule / Second 	Power

C 	Coulomb 	Second * Ampere 	Charge

V 	Volt 	Watt / Ampere 	ElectricPotential

F 	Farad 	Coulomb / Volt 	Capacitance

OHM 	Ohm 	Volt / Ampere 	Resistance

SIE 	Siemens 	Ampere / Volt 	Conductance

WB 	Weber 	Joule / Ampere 	MagneticFlux

T 	Tesla 	Weber / Meter2

H 	Henry 	Ohm * Second 	Inductance

LM 	Lumen 	Candela

LX 	Lux 	Candela / Meter2

BQ 	Becquerel 	Hertz

GY 	Gray 	Joule / Kilogram

SV 	Sievert 	Gray

KAT 	Katal 	Mole / Second

M2 	Meter2 	Meter * Meter 	Area

M3 	Meter3 	Meter2 * Meter 	Volume

PM 	PerMeter 	Unitless / Meter 	ReciprocalLength

PM2 	PerMeter2 	PerMeter / Meter

PM3 	PerMeter3 	PerMeter2 / Meter

S2 	Second2 	Second * Second

S3 	Second3 	Second2 * Second

S4 	Second4 	Second3 * Second

PS2 	PerSecond2 	Unitless / Second2

PS3 	PerSecond3 	Unitless / Second3

PS4 	PerSecond4 	Unitless / Second4

MPS 	MeterPerSecond 	Meter / Second 	Velocity

MPS2 	MeterPerSecond2 	Meter / Second2 	Acceleration

MPS3 	MeterPerSecond3 	Meter / Second3 	Jerk

MPS4 	MeterPerSecond4 	Meter / Second4

M2PS 	Meter2PerSecond 	Meter2 / Second

M2PS2 	Meter2PerSecond2 	Meter2 / Second2

M2PS3 	Meter2PerSecond3 	Meter2 / Second3

M3PS 	Meter3PerSecond 	Meter3 / Second

M3PS2 	Meter3PerSecond2 	Meter3 / Second2

M3PS3 	Meter3PerSecond3 	Meter3 / Second3

NS 	NewtonSecond 	Newton * Second

JS 	JouleSecond 	Joule * Second

NPS 	NewtonPerSecond 	Newton / Second

KGPM 	KilogramPerMeter 	Kilogram / Meter

KGPM2 	KilogramPerMeter2 	Kilogram / Meter2

KGPM3 	KilogramPerMeter3 	Kilogram / Meter3

MPKG 	MeterPerKilogram 	Meter / Kilogram

M2PKG 	Meter2PerKilogram 	Meter2 / Kilogram

M3PKG 	Meter3PerKilogram 	Meter3 / Kilogram

MOLPM 	MolePerMeter 	Mole / Meter

MOLPM2 	MolePerMeter2 	Mole / Meter2

MOLPM3 	MolePerMeter3 	Mole / Meter3

MPMOL 	MeterPerMole 	Meter / Mole

M2PMOL 	Meter2PerMole 	Meter2 / Mole

M3PMOL 	Meter3PerMole 	Meter3 / Mole

JPK 	JoulePerKelvin 	Joule / Kelvin

JPKMOL 	JoulePerKelvinMole 	JoulePerKelvin / Mole

JPKGK 	JoulePerKilogramKelvin 	JoulePerKelvin / Kilogram

JPMOL 	JoulePerMole 	Joule / Mole

JPKG 	JoulePerKilogram 	Joule / Kilogram

JPM2 	JoulePerMeter2 	Joule / Meter2

JPM3 	JoulePerMeter3 	Joule / Meter3

NPM 	NewtonPerMeter 	JoulePerMeter2

WPM2 	WattPerMeter2 	Watt / Meter2

WPMK 	WattPerMeterKelvin 	Watt / Meter / Kelvin

PAS 	PascalSecond 	Pascal * Second

CPM2 	CoulombPerMeter2 	Coulomb / Meter2

CPM3 	CoulombPerMeter3 	Coulomb / Meter3

APM2 	AmperePerMeter2 	Ampere / Meter2

SIEPM 	SiemensPerMeter 	Siemens / Meter

SIEM2PMOL 	SiemensMeter2PerMole 	Siemens * Meter2 / Mole

FPM 	FaradPerMeter 	Farad / Meter

HPM 	HenryPerMeter 	Henry / Meter

VPM 	VoltPerMeter 	Volt / Meter

APM 	AmperePerMeter 	Ampere / Meter

CDPM2 	CandelaPerMeter2 	Candela / Meter2

LMS 	LumenSecond 	Lumen * Second

LXS 	LuxSecond 	Lux * Second

CPKG 	CoulombPerKilogram 	Coulomb / Kilogram

GYPS 	GrayPerSecond 	Gray / Second

OHMM 	OhmMeter 	Ohm * Meter

CPM 	CoulombPerMeter 	Coulomb / Meter

MOLPKG 	MolePerKilogram 	Mole / Kilogram

KGPMOL 	KilogramPerMole 	Kilogram / Mole

KGPS 	KilogramPerSecomd 	Kilogram / Second

JPT 	JoulePerTesla 	Joule / Tesla

WPM3 	WattPerMeter3 	Watt / Meter3

KPW 	KelvinPerWatt 	Kelvin / Watt

PK 	PerKelvin 	Unitless / Kelvin

KPM 	KelvinPerMeter 	Kelvin / Meter

M2PVS 	Meter2PerVoltSecond 	Meter2 / Volt / Second

JPM2S 	JoulePerMeter2Second 	JoulePerMeter2 / Second

PPA 	PerPascal 	Unitless / Pascal

PH 	PerHenry 	Unitless / Henry

WBPM 	WeberPerMeter 	Weber / Meter

WBM 	WeberMeter 	Weber * Meter

TM 	TeslaMeter 	Tesla * Meter

M3PMOLS 	Meter3PerMoleSecond 	Meter3 / Mole / Second

NMSPKG 	NewtonMeterSecondPerKilogram 	Newton * Meter * Second / Kilogram

LMPW 	LumenPerWatt 	Lumen / Watt

MPH 	MeterPerHenry 	Meter / Henry

WPM 	WattPerMeter 	Watt / Meter

Constants

Name 	Constant 	Value 	Unit 	Dimension

Astronomical unit 	AU 	149597870700.0 * M 	Meter 	Length

Bohr radius 	R_BOHR 	0.52917721067e-10 * M 	Meter 	Length

Ångström 	AO 	1e-10 * M 	Meter 	Length

Foot 	FT 	0.3048 * M 	Meter 	Length

Inch 	IN 	FT / 12.0 	Meter 	Length

Yard 	YD 	3.0 * FT 	Meter 	Length

Mile 	MI 	5280.0 * FT 	Meter 	Length

Tonne 	TNE 	1.0e3 * KG 	Kilogram 	Mass

Unified atomic mass unit; dalton 	U 	1.660539040e-27 * KG 	Kilogram 	Mass

Electron mass 	M_E 	9.10938356e-31 * KG 	Kilogram 	Mass

Pound mass 	LB 	0.45359237 * KG 	Kilogram 	Mass

Ounce 	OZ 	LB / 16.0 	Kilogram 	Mass

Minute 	MIN 	60.0 * S 	Second 	Time

Hour 	HR 	60.0 * MIN 	Second 	Time

Day 	DAY 	24.0 * HR 	Second 	Time

Dyne 	DYN 	1.0e-5 * N 	Newton 	Force

Pound force 	LBF 	4.4482216152605 * N 	Newton 	Force

Bar 	BAR 	1e5 * PA 	Pascal 	Pressure

Millibar 	MBAR 	100.0 * PA 	Pascal 	Pressure

Atmosphere 	ATM 	101325.0 * PA 	Pascal 	Pressure

Barye 	BA 	0.1 * PA 	Pascal 	Pressure

Millimeter of mercury 	MMHG 	133.322387415 * PA 	Pascal 	Pressure

Torr 	TORR 	ATM / 760.0 	Pascal 	Pressure

Electronvolt 	EV 	E * V 	Joule 	Energy

Hartree energy 	EH 	4.359744650e-18 * J 	Joule 	Energy

Elementary charge 	E 	1.6021766208e-19 * C 	Coulomb 	Charge

Hectare 	HA 	10000.0 * M2 	Meter2 	Area

Are 	ARE 	100.0 * M2 	Meter2 	Area

Barn 	BARN 	1e-28 * M2 	Meter2 	Area

Liter 	L 	0.001 * M3 	Meter3 	Volume

Speed of light in a vacuum 	C0 	299792458.0 * MPS 	MeterPerSecond 	Velocity

Reduced Planck constant 	HBAR 	1.054571800e-34 * JS 	JouleSecond

Reexports

pub use self::f64consts::*;

Modules

f32consts
Constants defined for this system

f64consts
Constants defined for this system

Structs

SI
The struct for this unit system

Type Definitions

Ampere

AmperePerMeter

AmperePerMeter2

Becquerel

Candela

CandelaPerMeter2

Coulomb

CoulombPerKilogram

CoulombPerMeter

CoulombPerMeter2

CoulombPerMeter3

Farad

FaradPerMeter

Gray

GrayPerSecond

Henry

HenryPerMeter

Hertz

Joule

JoulePerKelvin

JoulePerKelvinMole

JoulePerKilogram

JoulePerKilogramKelvin

JoulePerMeter2

JoulePerMeter2Second

JoulePerMeter3

JoulePerMole

JoulePerTesla

JouleSecond

Katal

Kelvin

KelvinPerMeter

KelvinPerWatt

Kilogram

KilogramPerMeter

KilogramPerMeter2

KilogramPerMeter3

KilogramPerMole

KilogramPerSecomd

Lumen

LumenPerWatt

LumenSecond

Lux

LuxSecond

Meter

Meter2

Meter2PerKilogram

Meter2PerMole

Meter2PerSecond

Meter2PerSecond2

Meter2PerSecond3

Meter2PerVoltSecond

Meter3

Meter3PerKilogram

Meter3PerMole

Meter3PerMoleSecond

Meter3PerSecond

Meter3PerSecond2

Meter3PerSecond3

MeterPerHenry

MeterPerKilogram

MeterPerMole

MeterPerSecond

MeterPerSecond2

MeterPerSecond3

MeterPerSecond4

Mole

MolePerKilogram

MolePerMeter

MolePerMeter2

MolePerMeter3

Newton

NewtonMeterSecondPerKilogram

NewtonPerMeter

NewtonPerSecond

NewtonSecond

Ohm

OhmMeter

Pascal

PascalSecond

PerHenry

PerKelvin

PerMeter

PerMeter2

PerMeter3

PerPascal

PerSecond2

PerSecond3

PerSecond4

Second

Second2

Second3

Second4

Siemens

SiemensMeter2PerMole

SiemensPerMeter

Sievert

Tesla

TeslaMeter

Unitless

Volt

VoltPerMeter

Watt

WattPerMeter

WattPerMeter2

WattPerMeter3

WattPerMeterKelvin

Weber

WeberMeter

WeberPerMeter