Crate dimensioned

Reexports

pub use peano::*;

Modules

cgs

peano

si

Macros

__make_types!

dim_binary!

dim_lhs_div!
Scalar division (with scalar on LHS)!

dim_lhs_mult!
Scalar multiplication (with scalar on LHS)!

dim_unary!

make_units!

Structs

Dim

PhantomData
PhantomData<T> allows you to describe that a type acts as if it stores a value of type T, even though it does not. This allows you to inform the compiler about certain safety properties of your code.

Traits

AddDim

DimToString

Dimension

Dimensionless

DivDim

KeepDim

MulDim

NegDim

NotDim

Sqr

Sqrt

SubDim

Wrap

Implementations