Crate dimensioned

dimensioned

dimensioned is a library for compile time type checking for arbitrary unit systems.

The best place to read about it is here

Reexports

pub use dimensioned::*;

Modules

cgs
The cgs module provides a unit system for use with Gaussian CGS units.

dimensioned
This module allows dimensioned to be very flexible. It creates the Dim<D, V> type, which is the type that will be used with all dimensionful objects. It then implements as many traits from std as generically as possible.

peano
Peano numbers allow us to do arithmetic at compile time using Rust's type system.

si
The si module provides a unit system for use with SI units.

Macros

__make_types!

count_args!
Just counts the number of arguments its called with and gives you the total

dim_impl_binary!

dim_impl_unary!

make_units!

make_units_adv!