Module dimensioned::dimensioned

Reexports

pub use peano::*;

Structs

Dim

PhantomData
PhantomData is a way to tell the compiler about fake fields. Phantom data is required whenever type parameters are not used. The idea is that if the compiler encounters a PhantomData<T> instance, it will behave as if an instance of the type T were present for the purpose of various automatic analyses.

Traits

AddDim

DimToString

Dimension

Dimensionless

DivDim

KeepDim

MulDim

Scalar

Sqr

Sqrt

SubDim

Wrap